Use info_teleport_destination to determine stage zone angles
